+++ packages/ketcher-core/src/application/editor/EditorHistory.ts
@@ -15,7 +15,7 @@
  ***************************************************************************/
 
 import { Command } from 'domain/entities/Command';
-import type { CoreEditor } from './Editor';
+import { CoreEditor } from './Editor';
 import assert from 'assert';
 import { ketcherProvider } from 'application/utils';
 const HISTORY_SIZE = 32; // put me to options
