+++ packages/ketcher-core/src/application/editor/operations/rgroupAttachmentPoint/RGroupAttachmentPointAdd.ts
@@ -4,6 +4,7 @@ import {
   RGroupAttachmentPoint,
   RGroupAttachmentPointType,
 } from 'domain/entities';
+import { RGroupAttachmentPointRemove } from '.';
 import { OperationPriority, OperationType } from '../OperationType';
 import BaseOperation from '../base';
 
@@ -55,9 +56,6 @@ class RGroupAttachmentPointAdd extends BaseOperation {
     if (this.data.attachmentPointId === undefined) {
       throw Error(`Inverted attachmentPointId doesn't exist`);
     }
-    const {
-      RGroupAttachmentPointRemove,
-    } = require('./RGroupAttachmentPointRemove');
     return new RGroupAttachmentPointRemove(this.data.attachmentPointId);
   }
 }
