+++ packages/ketcher-core/src/application/editor/operations/rgroupAttachmentPoint/RGroupAttachmentPointRemove.ts
@@ -1,5 +1,6 @@
 import { ReStruct } from 'application/render';
 import assert from 'assert';
+import { RGroupAttachmentPointAdd } from '.';
 import { OperationPriority, OperationType } from '../OperationType';
 import BaseOperation from '../base';
 
@@ -43,7 +44,6 @@ class RGroupAttachmentPointRemove extends BaseOperation {
   }
 
   invert() {
-    const { RGroupAttachmentPointAdd } = require('./RGroupAttachmentPointAdd');
     return new RGroupAttachmentPointAdd(this.data);
   }
 }
