+++ packages/ketcher-react/src/script/ui/state/templates/init-lib.ts
@@ -22,6 +22,7 @@ import {
   SettingsManager,
 } from 'ketcher-core';
 
+import { appUpdate } from '../options';
 import { storage } from '../../storage-ext';
 import templatesRawData from '../../../../templates/library.sdf';
 import { AnyAction, Dispatch } from 'redux';
@@ -78,8 +79,7 @@ export default async function initTmplLib(
   return deserializeSdfTemplates(baseUrl, cacheEl, fileName).then((res) => {
     const lib = res.concat(userTmpls());
     dispatch(initLib(lib));
-    dispatch({ type: 'APP_OPTIONS', data: { templates: true } });
-    dispatch({ type: 'UPDATE' });
+    dispatch(appUpdate({ templates: true }) as unknown as AnyAction);
     if (needReinitializeTemplateLibrary) {
       reinitializeTemplateLibrary();
       needReinitializeTemplateLibrary = false;
