--- packages/ketcher-react/src/script/ui/views/modal/components/ExtendedTable/components/GenericGroups/groupNames.ts
@@ -1,11 +1,0 @@
-export enum groupNames {
-  atomsGen = 'atoms-gen',
-  specialNodes = 'special-nodes',
-  groupGen = 'group-gen',
-  groupAcyclic = 'group-acyclic',
-  acyclicCarbo = 'acyclic-carbo',
-  acyclicHetero = 'acyclic-hetero',
-  groupCyclic = 'group-cyclic',
-  cyclicHetero = 'cyclic-hetero',
-  cyclicCarbo = 'cyclic-carbo',
-}
