+++ packages/ketcher-react/src/script/ui/views/modal/components/ExtendedTable/components/GenericGroups/index.ts
@@ -15,4 +15,15 @@
  ***************************************************************************/
 
 export { default } from './GenericGroups';
-export { groupNames } from './groupNames';
+
+export enum groupNames {
+  atomsGen = 'atoms-gen',
+  specialNodes = 'special-nodes',
+  groupGen = 'group-gen',
+  groupAcyclic = 'group-acyclic',
+  acyclicCarbo = 'acyclic-carbo',
+  acyclicHetero = 'acyclic-hetero',
+  groupCyclic = 'group-cyclic',
+  cyclicHetero = 'cyclic-hetero',
+  cyclicCarbo = 'cyclic-carbo',
+}
